+++ cmd/notificationsrocketchat.go
@@ -105,15 +105,19 @@ var deleteProjectRocketChatNotificationCmd = &cobra.Command{
 			cmd.Help()
 			os.Exit(1)
 		}
-		deleteResult, err := pClient.DeleteRocketChatNotificationFromProject(notificationFlags.Project, notificationFlags.NotificationName)
-		handleError(err)
-		var addedProject api.NotificationSlack
-		err = json.Unmarshal([]byte(deleteResult), &addedProject)
-		handleError(err)
-		resultData := output.Result{
-			Result: "success",
+		fmt.Println(fmt.Sprintf("Deleting notification %s from project %s", notificationFlags.NotificationName, notificationFlags.Project))
+
+		if yesNo("Are you sure you want to delete?") {
+			deleteResult, err := pClient.DeleteRocketChatNotificationFromProject(notificationFlags.Project, notificationFlags.NotificationName)
+			handleError(err)
+			var addedProject api.NotificationSlack
+			err = json.Unmarshal([]byte(deleteResult), &addedProject)
+			handleError(err)
+			resultData := output.Result{
+				Result: "success",
+			}
+			output.RenderResult(resultData, outputOptions)
 		}
-		output.RenderResult(resultData, outputOptions)
 	},
 }
 var deleteRocketChatNotificationCmd = &cobra.Command{
@@ -127,12 +131,16 @@ var deleteRocketChatNotificationCmd = &cobra.Command{
 			cmd.Help()
 			os.Exit(1)
 		}
-		deleteResult, err := pClient.DeleteRocketChatNotification(notificationFlags.NotificationName)
-		handleError(err)
-		resultData := output.Result{
-			Result: string(deleteResult),
+		fmt.Println(fmt.Sprintf("Deleting notification %s", notificationFlags.NotificationName))
+
+		if yesNo("Are you sure you want to delete?") {
+			deleteResult, err := pClient.DeleteRocketChatNotification(notificationFlags.NotificationName)
+			handleError(err)
+			resultData := output.Result{
+				Result: string(deleteResult),
+			}
+			output.RenderResult(resultData, outputOptions)
 		}
-		output.RenderResult(resultData, outputOptions)
 	},
 }
