+++ cmd/users.go
@@ -118,12 +118,16 @@ var deleteSSHKeyCmd = &cobra.Command{
 		}
 		var customReqResult []byte
 		var err error
-		customReqResult, err = uClient.DeleteSSHKey(sshKeyName)
-		handleError(err)
-		resultData := output.Result{
-			Result: string(customReqResult),
+		fmt.Println(fmt.Sprintf("Deleting ssh key %s", sshKeyName))
+
+		if yesNo("Are you sure you want to delete?") {
+			customReqResult, err = uClient.DeleteSSHKey(sshKeyName)
+			handleError(err)
+			resultData := output.Result{
+				Result: string(customReqResult),
+			}
+			output.RenderResult(resultData, outputOptions)
 		}
-		output.RenderResult(resultData, outputOptions)
 	},
 }
 
@@ -140,12 +144,16 @@ var deleteUserCmd = &cobra.Command{
 		}
 		var customReqResult []byte
 		var err error
-		customReqResult, err = uClient.DeleteUser(userFlags)
-		handleError(err)
-		resultData := output.Result{
-			Result: string(customReqResult),
+		fmt.Println(fmt.Sprintf("Deleting user with email %s", userFlags.Email))
+
+		if yesNo("Are you sure you want to delete?") {
+			customReqResult, err = uClient.DeleteUser(userFlags)
+			handleError(err)
+			resultData := output.Result{
+				Result: string(customReqResult),
+			}
+			output.RenderResult(resultData, outputOptions)
 		}
-		output.RenderResult(resultData, outputOptions)
 	},
 }
